[BE] Add checksum situation
Currently uses a canonical form of JSON (sorted fields and no spacing),
adler32'd (which is apparently a decent checksum to use in this sort of
case). I'll implement client-side tomorrow. Untested as yet, I just
wanted to commit my work for tonight. The details may change if client
implementation proves an issue.